# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PawZone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up as buyer, seller or transporter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/waitlist:
    post:
      operationId: addToWaitlist
      tags: [auth]
      summary: Add user to waitlist for non-Kerala region
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistBody"
      responses:
        "201":
          description: Added to waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Get all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/approve:
    post:
      operationId: adminApproveUser
      tags: [admin]
      summary: Approve a seller or transporter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/block:
    post:
      operationId: adminBlockUser
      tags: [admin]
      summary: Block a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/listings:
    get:
      operationId: adminGetListings
      tags: [admin]
      summary: Get all listings for admin review
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /admin/listings/{id}/approve:
    post:
      operationId: adminApproveListing
      tags: [admin]
      summary: Approve a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /admin/listings/{id}/reject:
    post:
      operationId: adminRejectListing
      tags: [admin]
      summary: Reject a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectBody"
      responses:
        "200":
          description: Listing rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
  /admin/disputes/{id}/resolve:
    post:
      operationId: adminResolveDispute
      tags: [admin]
      summary: Resolve a dispute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveDisputeBody"
      responses:
        "200":
          description: Dispute resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
  /admin/refunds/{orderId}/approve:
    post:
      operationId: adminApproveRefund
      tags: [admin]
      summary: Approve a refund
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Refund approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginBody:
      type: object
      properties:
        loginId:
          type: string
        password:
          type: string
      required:
        - loginId
        - password
    SignupBody:
      type: object
      properties:
        role:
          type: string
          enum: [buyer, seller, transporter]
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required:
        - role
        - name
        - email
        - phone
        - password
        - address
        - city
        - state
        - pincode
        - country
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    WaitlistBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
      required:
        - name
        - email
        - city
        - state
        - country
...
    User:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [buyer, seller, transporter, admin]
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [pending, approved, blocked]
        sellerId:
          type: string
          nullable: true
        sellerScore:
          type: number
          nullable: true
        profilePhoto:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - name
        - email
        - status
        - createdAt
    Listing:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        category:
          type: string
          enum: [dogs, cats, birds, fish]
        breed:
          type: string
        price:
          type: number
        quantity:
          type: integer
        availableQuantity:
          type: integer
        vaccinated:
          type: boolean
        vaccinationDetails:
          type: string
          nullable: true
        photos:
          type: array
          items:
            type: string
        videoUrl:
          type: string
          nullable: true
        description:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, sold_out]
        petCode:
          type: string
          nullable: true
        address:
          type: string
        city:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sellerId
        - sellerName
        - category
        - breed
        - price
        - quantity
        - availableQuantity
        - vaccinated
        - photos
        - description
        - status
        - address
        - city
        - createdAt
...
    Dispute:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        orderNumber:
          type: string
        reportedBy:
          type: integer
        reportedByName:
          type: string
        issueType:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_review, resolved, closed]
        resolution:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - orderId
        - orderNumber
        - reportedBy
        - reportedByName
        - issueType
        - description
        - status
        - createdAt
    ResolveDisputeBody:
      type: object
      properties:
        resolution:
          type: string
        action:
          type: string
          enum: [refund, no_action, warn_seller, warn_buyer]
      required:
        - resolution
        - action
    RejectBody:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
...